Version 2.3 - Added ability to append the current date and/or the file version to the output filename - Updated user guide

+++ sfx/sfxPackager/GenParser.cpp
@@ -31,7 +31,7 @@ CGenParser::~CGenParser()
 }
 
 
-void CGenParser::SetSourceData(const TCHAR *data, UINT datalen)
+void CGenParser::SetSourceData(const TCHAR *data, size_t datalen)
 {
 	m_data = (TCHAR *)data;
 	m_datalen = datalen;
